+++ src/Voting.Stimmunterlagen.Data/Models/VotingCardLayout.cs
@@ -18,6 +18,4 @@ public abstract class VotingCardLayout : BaseEntity
     public VotingCardLayoutDataConfiguration DataConfiguration { get; set; } = new();
 
     public DomainOfInfluenceVotingCardPrintData? PrintData { get; set; }
-
-    public VotingCardColor VotingCardColor { get; set; }
 }