+++ src/Voting.Stimmunterlagen/MappingProfiles/ContestVotingCardLayoutProfile.cs
@@ -11,8 +11,7 @@ public class ContestVotingCardLayoutProfile : Profile
 {
     public ContestVotingCardLayoutProfile()
     {
-        CreateMap<Data.Models.ContestVotingCardLayout, ProtoModels.ContestVotingCardLayout>()
-            .ForMember(dst => dst.Color, opts => opts.MapFrom(x => x.VotingCardColor));
+        CreateMap<Data.Models.ContestVotingCardLayout, ProtoModels.ContestVotingCardLayout>();
         CreateMap<IEnumerable<Data.Models.ContestVotingCardLayout>, ProtoModels.ContestVotingCardLayouts>()
             .ForMember(dst => dst.Layouts, opts => opts.MapFrom(x => x));
     }