+++ src/Voting.Stimmunterlagen/MappingProfiles/EnumProfile.cs
@@ -5,7 +5,6 @@ using System;
 using AutoMapper;
 using AutoMapper.Extensions.EnumMapping;
 using Voting.Stimmunterlagen.Data.Models;
-using BasisProtoModels = Abraxas.Voting.Basis.Shared.V1;
 using ProtoModels = Voting.Stimmunterlagen.Proto.V1.Models;
 
 namespace Voting.Stimmunterlagen.MappingProfiles;
@@ -29,14 +28,14 @@ public class EnumProfile : Profile
         CreateEnumMap<VotingCardSort, ProtoModels.VotingCardSort>();
         CreateEnumMap<PrintJobState, ProtoModels.PrintJobState>();
         CreateEnumMap<VotingCardGeneratorJobState, ProtoModels.VotingCardGeneratorJobState>();
-        CreateEnumMap<VotingCardColor, ProtoModels.VotingCardColor>();
 
         // explicitly map deprecated values to default value.
-        CreateMap<BasisProtoModels.VotingCardColor, VotingCardColor>()
+        CreateMap<Abraxas.Voting.Basis.Shared.V1.VotingCardColor, VotingCardColor>()
             .ConvertUsingEnumMapping(opt => opt
                 .MapByValue()
-                .MapValue(BasisProtoModels.VotingCardColor.Chamois, VotingCardColor.Unspecified)
-                .MapValue(BasisProtoModels.VotingCardColor.Gold, VotingCardColor.Unspecified));
+                .MapValue(Abraxas.Voting.Basis.Shared.V1.VotingCardColor.Chamois, VotingCardColor.Unspecified)
+                .MapValue(Abraxas.Voting.Basis.Shared.V1.VotingCardColor.Gold, VotingCardColor.Unspecified))
+            .ReverseMap();
     }
 
     private void CreateEnumMap<T1, T2>()